Refactor to convert values to float in browser
This gets things closer to what I ultimately want, but still needs work.
It now stores the metric value as a string, and converts it to a float
in the browser. What I am trying to do is to keep things simple, both in
code and for the user. So storing values without requiring the user to
do a lot of esoteric configuration (i.e. define value datatype,
precision, scale, etc. and be able to convert and display using the
user's browser locale.