Rename columns and fix date handling
This renames the awkard goal_timestamp to goal_date and display_units to
units. It also makes description optional.

Most importantly, it fixes the conversion of dates so that the html
inputs can handle and display them correctly. I may yet have timezone
issues, so I need to test that.