Add new cicd workflow and post deploy.sh script
I'm abandoning my attempt at independent test/build/deploy workflows.
GitHub Actions is just not mature enough to work this way. The current
commit combines them into a single new cicd.yml workflow and dissables
the others. It is still not an ideal CI/CD solution but seems to be
adequate for now.